Fix a warning

+++ src/properties/trackerlist.cpp
@@ -427,7 +427,7 @@ void TrackerList::reannounceSelected() {
       return;
 
     std::vector<announce_entry> trackers = h.trackers();
-    for (int i = 0; i < trackers.size(); ++i) {
+    for (size_t i = 0; i < trackers.size(); ++i) {
       foreach (QTreeWidgetItem* w, selected_items) {
         if (w->text(COL_URL) == misc::toQString(trackers[i].url)) {
           h.force_reannounce(0, i);
